#59 [REFACTOR] 계정 찾기 로직을 훅으로 분리
